+++ sai-challenge11/App.test.js
@@ -4,6 +4,6 @@ import App from "./App";
 import renderer from "react-test-renderer";
 
 it("renders without crashing", () => {
-  const rendered = renderer.create(<Home />).toJSON();
+  const rendered = renderer.create(<App />).toJSON();
   expect(rendered).toBeTruthy();
 });
